rename : 파일명 및 디렉토리 구조 수정
userInfo.java -> FindUserInfo.java
utils 디렉토리 생성 후 이미지 업로드 관련 클래스 넣어줌
