+++ src/main/java/dev/umc/healody/common/FindUserInfo.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 
 @RequiredArgsConstructor
-public class userInfo {
+public class FindUserInfo {
 
     public static UserService userService;
