+++ src/main/java/dev/umc/healody/family/FamilyApiController.java
@@ -4,12 +4,7 @@ import dev.umc.healody.common.SuccessResponse;
 import dev.umc.healody.common.SuccessStatus;
 import dev.umc.healody.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-
-import static dev.umc.healody.common.userInfo.getCurrentUserId;
 
 @RequiredArgsConstructor
 @RestController
