+++ src/main/java/dev/umc/healody/utils/S3Config.java
@@ -1,4 +1,4 @@
-package dev.umc.healody.common;
+package dev.umc.healody.utils;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
